Added a User Service to interact with DB
In the last commit, forgot to commit the UserService class. As
previously mention, created a basic login and signup methods for a
controller to use. Methods were also created to insure users cannot have
the same username or email. User class added a few functions to hash
passwords and check if a password is equal to that hash password for the
hashing password commit.